% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_frame.R
\name{as_sparse_annotation_factor}
\alias{as_sparse_annotation_factor}
\title{Convert a categorical annotation to a sparse-aware factor}
\usage{
as_sparse_annotation_factor(
  x,
  column,
  control,
  preprocessing,
  collapse_sparse = TRUE
)
}
\description{
Convert a categorical annotation to a sparse-aware factor
}
\keyword{internal}
