% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_spec.R
\name{assay_feature}
\alias{assay_feature}
\title{Define an assay feature response}
\usage{
assay_feature(assay, feature)
}
\arguments{
\item{assay}{Name of the assay in a `LongitudinalDataset`.}

\item{feature}{Name of the feature within the assay.}
}
\value{
A LEVAiM response object.
}
\description{
Define an assay feature response
}
