% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data_io-io_humann.R
\name{clean_humann_sample_names}
\alias{clean_humann_sample_names}
\title{Clean HUMAnN sample names}
\usage{
clean_humann_sample_names(x)
}
\description{
Clean HUMAnN sample names
}
\keyword{internal}
