% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-compilation_engines.R
\name{compile_covariate_terms_gp}
\alias{compile_covariate_terms_gp}
\title{Compile covariate terms for GP/brms models}
\usage{
compile_covariate_terms_gp(covariates)
}
\description{
Compile covariate terms for GP/brms models
}
\keyword{internal}
