% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-compilation_engines.R
\name{compile_covariate_terms_spline}
\alias{compile_covariate_terms_spline}
\title{Compile covariate terms for spline/GAM models}
\usage{
compile_covariate_terms_spline(covariates)
}
\description{
Compile covariate terms for spline/GAM models
}
\keyword{internal}
