% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-compilation_engines.R
\name{compile_formula_spline}
\alias{compile_formula_spline}
\title{Compile a spline/GAM formula}
\usage{
compile_formula_spline(traj, k = NULL)
}
\arguments{
\item{traj}{A `levaim_trajectory` object.}

\item{k}{Optional basis dimension for spline smooths.}
}
\value{
A formula.
}
\description{
Compiles a LEVAiM trajectory into an `mgcv::gam()`-style formula.
}
\keyword{internal}
