% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_spec.R
\name{continuous_effect}
\alias{continuous_effect}
\title{Define a continuous model effect}
\usage{
continuous_effect(column, scale = FALSE)
}
\arguments{
\item{column}{Metadata column encoding a continuous covariate.}

\item{scale}{Logical. Whether to scale the covariate before modeling.}
}
\value{
A LEVAiM covariate object.
}
\description{
Define a continuous model effect
}
