% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_frame.R
\name{covariate_role_diagnostics}
\alias{covariate_role_diagnostics}
\title{Infer and validate covariate roles}
\usage{
covariate_role_diagnostics(data, traj)
}
\description{
Infer and validate covariate roles
}
\keyword{internal}
