% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-validation-cross_validation.R
\name{cv_summary}
\alias{cv_summary}
\title{Summarize CV metrics}
\usage{
cv_summary(x)
}
\arguments{
\item{x}{A `levaim_cv` object.}
}
\value{
A data.frame of aggregate CV metrics.
}
\description{
Summarize CV metrics
}
