% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data_io-io_humann.R
\name{detect_humann_feature_space}
\alias{detect_humann_feature_space}
\title{Detect HUMAnN feature space}
\usage{
detect_humann_feature_space(feature_ids, path = NULL)
}
\description{
Detect HUMAnN feature space
}
\keyword{internal}
