% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data_io-io_humann.R
\name{detect_humann_input_type}
\alias{detect_humann_input_type}
\title{Detect whether HUMAnN input is single-profile or merged}
\usage{
detect_humann_input_type(path)
}
\description{
Detect whether HUMAnN input is single-profile or merged
}
\keyword{internal}
