% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data_io-io_humann.R
\name{detect_humann_units}
\alias{detect_humann_units}
\title{Detect HUMAnN units}
\usage{
detect_humann_units(path = NULL, header = character())
}
\description{
Detect HUMAnN units
}
\keyword{internal}
