% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-preprocess-exposure_derivation.R
\name{exposure_derivations}
\alias{exposure_derivations}
\title{Return the exposure derivation audit table}
\usage{
exposure_derivations(ds)
}
\arguments{
\item{ds}{A `LongitudinalDataset`.}
}
\value{
A data.frame.
}
\description{
Return the exposure derivation audit table
}
