% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-preprocess-exposure_derivation.R
\name{exposure_ever}
\alias{exposure_ever}
\title{Derive whether an exposure value ever occurs}
\usage{
exposure_ever(value)
}
\arguments{
\item{value}{Exposure value to query.}
}
\value{
An exposure summary specification.
}
\description{
Derive whether an exposure value ever occurs
}
