% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-preprocess-exposure_derivation.R
\name{exposure_last_observed}
\alias{exposure_last_observed}
\title{Derive the last observed exposure value}
\usage{
exposure_last_observed()
}
\value{
An exposure summary specification.
}
\description{
Derive the last observed exposure value
}
