% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-preprocess-exposure_derivation.R
\name{exposure_transition}
\alias{exposure_transition}
\title{Derive the observed exposure transition path}
\usage{
exposure_transition()
}
\value{
An exposure summary specification.
}
\description{
Derive the observed exposure transition path
}
