% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_frame.R
\name{extract_response}
\alias{extract_response}
\title{Extract response vector from a LongitudinalDataset}
\usage{
extract_response(ds, response)
}
\description{
Extract response vector from a LongitudinalDataset
}
\keyword{internal}
