% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data_io-io_humann.R
\name{filter_humann_features}
\alias{filter_humann_features}
\title{Filter HUMAnN rows by stratification and special rows}
\usage{
filter_humann_features(
  feature_ids,
  stratification = c("community", "stratified", "all"),
  remove_unmapped = TRUE
)
}
\description{
Filter HUMAnN rows by stratification and special rows
}
\keyword{internal}
