% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-fit-fit_trajectory.R
\name{fit_trajectory}
\alias{fit_trajectory}
\title{Fit a LEVAiM trajectory}
\usage{
fit_trajectory(mf)
}
\arguments{
\item{mf}{A `levaim_model_frame`.}
}
\value{
A `levaim_fit`.
}
\description{
Fits a trajectory model using the backend specified by the associated
trajectory control object.
}
