% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-fit-fit_trajectory.R
\name{fit_trajectory_spline}
\alias{fit_trajectory_spline}
\title{Fit a spline trajectory}
\usage{
fit_trajectory_spline(mf)
}
\arguments{
\item{mf}{A `levaim_model_frame`.}
}
\value{
A `levaim_fit`.
}
\description{
Fit a spline trajectory
}
\keyword{internal}
