% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_spec.R
\name{group_effect}
\alias{group_effect}
\title{Define a categorical model effect}
\usage{
group_effect(column, reference = NULL)
}
\arguments{
\item{column}{Metadata column encoding a categorical/grouping effect.}

\item{reference}{Optional reference level.}
}
\value{
A LEVAiM covariate object.
}
\description{
Define a categorical model effect
}
