% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-serialization.R
\name{is_levaim_serializable}
\alias{is_levaim_serializable}
\title{Check whether an object uses a serializable LEVAiM class}
\usage{
is_levaim_serializable(x)
}
\arguments{
\item{x}{An object.}
}
\value{
`TRUE` when `x` has a recognized LEVAiM class.
}
\description{
Check whether an object uses a serializable LEVAiM class
}
\keyword{internal}
