% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_spec.R
\name{metadata_feature}
\alias{metadata_feature}
\title{Define a metadata feature response}
\usage{
metadata_feature(column)
}
\arguments{
\item{column}{Name of the metadata column.}
}
\value{
A LEVAiM response object.
}
\description{
Define a metadata feature response
}
