% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_spec.R
\name{model_spec}
\alias{model_spec}
\title{Define a longitudinal model specification}
\usage{
model_spec(response, subject, time, covariates = list())
}
\arguments{
\item{response}{Response definition, usually from `assay_feature()` or
`metadata_feature()`.}

\item{subject}{Subject identifier, from `subject_var()`.}

\item{time}{Chronology variable, from `time_var()`.}

\item{covariates}{List of covariate definitions.}
}
\value{
A `levaim_model_spec` object.
}
\description{
Declares the response, subject identifier, chronology variable, and model
covariates. This object does not define whether trajectories are shared or
vary by covariates; that is handled later by `trajectory()`.
}
