% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-trajectory_design.R
\name{naive_trajectory}
\alias{naive_trajectory}
\title{Define a naive (non-varying) trajectory}
\usage{
naive_trajectory()
}
\value{
A LEVAiM trajectory design object.
}
\description{
A naive trajectory uses a single time function across all samples.
}
