% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-plotting-trajectory_reports.R
\name{plot_trajectory_moments}
\alias{plot_trajectory_moments}
\title{Plot trajectory moment estimates}
\usage{
plot_trajectory_moments(
  moments,
  moment = c("mean_level", "auc", "amplitude", "linear_trend"),
  window = NULL,
  feature = NULL,
  palette = NULL
)
}
\arguments{
\item{moments}{Output from [trajectory_moments()].}

\item{moment}{Character vector of moments to display.}

\item{window}{Optional window labels to display.}

\item{feature}{Optional feature names to display.}

\item{palette}{Optional group color vector.}
}
\value{
A faceted `ggplot` object.
}
\description{
Plot trajectory moment estimates
}
