% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_frame.R
\name{preprocess_model_frame_annotations}
\alias{preprocess_model_frame_annotations}
\title{Preprocess sparse and missing model-frame annotations}
\usage{
preprocess_model_frame_annotations(data, traj, control)
}
\description{
Preprocess sparse and missing model-frame annotations
}
\keyword{internal}
