% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-validation-cross_validation.R
\name{print.levaim_cv}
\alias{print.levaim_cv}
\title{Print grouped cross-validation results}
\usage{
\method{print}{levaim_cv}(x, ...)
}
\arguments{
\item{x}{A `levaim_cv` object.}

\item{...}{Unused.}
}
\value{
Invisibly returns `x`.
}
\description{
Print grouped cross-validation results
}
