% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-workflows-feature_tests.R
\name{print.levaim_feature_tests}
\alias{print.levaim_feature_tests}
\title{Print LEVAiM feature tests}
\usage{
\method{print}{levaim_feature_tests}(x, ...)
}
\arguments{
\item{x}{A `levaim_feature_tests` object.}

\item{...}{Unused.}
}
\value{
Invisibly returns `x`.
}
\description{
Print LEVAiM feature tests
}
