% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_frame.R
\name{print.levaim_model_frame}
\alias{print.levaim_model_frame}
\title{Print a LEVAiM model frame}
\usage{
\method{print}{levaim_model_frame}(x, ...)
}
\arguments{
\item{x}{A `levaim_model_frame`.}

\item{...}{Unused.}
}
\value{
Invisibly returns `x`.
}
\description{
Print a LEVAiM model frame
}
