% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_spec.R
\name{print.levaim_model_spec}
\alias{print.levaim_model_spec}
\title{Print a LEVAiM model specification}
\usage{
\method{print}{levaim_model_spec}(x, ...)
}
\arguments{
\item{x}{A `levaim_model_spec`.}

\item{...}{Unused.}
}
\value{
Invisibly returns `x`.
}
\description{
Print a LEVAiM model specification
}
