% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-results-result_summary.R
\name{print.levaim_results}
\alias{print.levaim_results}
\title{Print LEVAiM results}
\usage{
\method{print}{levaim_results}(x, ...)
}
\arguments{
\item{x}{A `levaim_results` object.}

\item{...}{Unused.}
}
\value{
Invisibly returns `x`.
}
\description{
Print LEVAiM results
}
