% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-trajectory_control.R
\name{print.levaim_trajectory_control}
\alias{print.levaim_trajectory_control}
\title{Print trajectory modeling controls}
\usage{
\method{print}{levaim_trajectory_control}(x, ...)
}
\arguments{
\item{x}{A `levaim_trajectory_control` object.}

\item{...}{Unused.}
}
\value{
Invisibly returns `x`.
}
\description{
Print trajectory modeling controls
}
