% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes-LongitudinalDataset.R
\name{print.LongitudinalDataset}
\alias{print.LongitudinalDataset}
\title{Print a LongitudinalDataset}
\usage{
\method{print}{LongitudinalDataset}(x, ...)
}
\arguments{
\item{x}{A `LongitudinalDataset`.}

\item{...}{Unused.}
}
\value{
Invisibly returns `x`.
}
\description{
Print a LongitudinalDataset
}
