% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_spec.R
\name{random_effect}
\alias{random_effect}
\title{Define a random effect}
\usage{
random_effect(column)
}
\arguments{
\item{column}{Metadata column encoding a random-effect grouping variable.}
}
\value{
A LEVAiM covariate object.
}
\description{
Define a random effect
}
