% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data_io-io_humann.R
\name{read_humann_many}
\alias{read_humann_many}
\title{Read many single HUMAnN profiles}
\usage{
read_humann_many(
  paths,
  feature_space = c("auto", "uniref", "ko", "ec", "pathway", "genefamily", "other"),
  units = c("auto", "cpm", "rpk", "relative_abundance", "coverage", "other"),
  stratification = c("community", "stratified", "all"),
  remove_unmapped = TRUE,
  sample_names = NULL,
  clean_sample_names = TRUE
)
}
\description{
Read many single HUMAnN profiles
}
\keyword{internal}
