% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data_io-io_humann.R
\name{read_humann_merged}
\alias{read_humann_merged}
\title{Read a merged HUMAnN table}
\usage{
read_humann_merged(
  path,
  feature_space = c("auto", "uniref", "ko", "ec", "pathway", "genefamily", "other"),
  units = c("auto", "cpm", "rpk", "relative_abundance", "coverage", "other"),
  stratification = c("community", "stratified", "all"),
  remove_unmapped = TRUE,
  clean_sample_names = TRUE
)
}
\description{
Read a merged HUMAnN table
}
\keyword{internal}
