% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data_io-io_humann.R
\name{read_humann_single}
\alias{read_humann_single}
\title{Read a single HUMAnN profile}
\usage{
read_humann_single(
  path,
  feature_space = c("auto", "uniref", "ko", "ec", "pathway", "genefamily", "other"),
  units = c("auto", "cpm", "rpk", "relative_abundance", "coverage", "other"),
  stratification = c("community", "stratified", "all"),
  remove_unmapped = TRUE,
  sample_name = NULL,
  clean_sample_names = TRUE
)
}
\description{
Read a single HUMAnN profile
}
\keyword{internal}
