% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-data_io-io_humann.R
\name{read_humann_table_fast}
\alias{read_humann_table_fast}
\title{Fast-read a HUMAnN-style table}
\usage{
read_humann_table_fast(path)
}
\description{
Fast-read a HUMAnN-style table
}
\keyword{internal}
