% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_spec.R
\name{subject_var}
\alias{subject_var}
\title{Define the subject identifier}
\usage{
subject_var(column)
}
\arguments{
\item{column}{Metadata column identifying subjects.}
}
\value{
A LEVAiM subject variable.
}
\description{
Define the subject identifier
}
