% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_spec.R
\name{time_var}
\alias{time_var}
\title{Define the time variable}
\usage{
time_var(column)
}
\arguments{
\item{column}{Metadata column encoding chronology.}
}
\value{
A LEVAiM time variable.
}
\description{
Define the time variable
}
