% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-results-result_summary.R
\name{trajectory_results_gp}
\alias{trajectory_results_gp}
\title{Extract Gaussian process results}
\usage{
trajectory_results_gp(fit)
}
\description{
Extract Gaussian process results
}
\keyword{internal}
