% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-results-result_summary.R
\name{trajectory_results_spline}
\alias{trajectory_results_spline}
\title{Extract spline/GAM results}
\usage{
trajectory_results_spline(fit)
}
\description{
Extract spline/GAM results
}
\keyword{internal}
