% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-classes-LongitudinalDataset.R
\name{validate_longitudinal_dataset}
\alias{validate_longitudinal_dataset}
\title{Validate a longitudinal multi-assay dataset}
\usage{
validate_longitudinal_dataset(assays, metadata, time_col, subject_col)
}
\arguments{
\item{assays}{Named list of assay tables.}

\item{metadata}{Sample metadata table.}

\item{time_col}{Metadata column encoding time.}

\item{subject_col}{Metadata column encoding subject ID.}
}
\value{
Invisibly returns `TRUE`.
}
\description{
Validate a longitudinal multi-assay dataset
}
\keyword{internal}
