% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_frame.R
\name{validate_model_frame_columns}
\alias{validate_model_frame_columns}
\title{Validate model frame columns required by trajectory}
\usage{
validate_model_frame_columns(mf, traj)
}
\description{
Validate model frame columns required by trajectory
}
\keyword{internal}
