% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_spec.R
\name{validate_model_spec_inputs}
\alias{validate_model_spec_inputs}
\title{Validate model specification inputs}
\usage{
validate_model_spec_inputs(response, subject, time, covariates)
}
\description{
Validate model specification inputs
}
\keyword{internal}
