% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-model_frame.R
\name{validate_response_family}
\alias{validate_response_family}
\title{Validate the response domain for a model family}
\usage{
validate_response_family(y, family)
}
\description{
Validate the response domain for a model family
}
\keyword{internal}
