% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-trajectory_design.R
\name{validate_trajectory_design}
\alias{validate_trajectory_design}
\title{Validate a trajectory design against a model specification}
\usage{
validate_trajectory_design(spec, design)
}
\description{
Validate a trajectory design against a model specification
}
\keyword{internal}
