% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-trajectory_design.R
\name{varying_trajectory}
\alias{varying_trajectory}
\title{Define a covariate-varying trajectory}
\usage{
varying_trajectory(by, interaction = length(by) > 1L)
}
\arguments{
\item{by}{Character vector of covariate column names.}

\item{interaction}{Logical. If `TRUE`, multiple `by` variables are combined
into a single interaction factor.}
}
\value{
A LEVAiM trajectory design object.
}
\description{
A varying trajectory allows the time function to differ across levels of
one or more declared covariates.
}
