% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-compilation_engines.R
\name{compile_formula}
\alias{compile_formula}
\title{Compile a trajectory object into a model formula}
\usage{
compile_formula(traj, engine = c("spline", "gp"))
}
\arguments{
\item{traj}{A `levaim_trajectory` object.}

\item{engine}{Modeling engine. One of `"spline"` or `"gp"`.}
}
\value{
A formula.
}
\description{
Compile a trajectory object into a model formula
}
\keyword{internal}
