% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-compilation_engines.R
\name{compile_formula_gp}
\alias{compile_formula_gp}
\title{Compile a GP formula}
\usage{
compile_formula_gp(traj)
}
\arguments{
\item{traj}{A `levaim_trajectory` object.}
}
\value{
A formula.
}
\description{
Compiles a LEVAiM trajectory into a `brms::brm()`-style formula.
}
\keyword{internal}
