% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-compilation_engines.R
\name{compile_time_terms_gp}
\alias{compile_time_terms_gp}
\title{Compile time trajectory terms for GP/brms models}
\usage{
compile_time_terms_gp(time_col, design)
}
\description{
Compile time trajectory terms for GP/brms models
}
\keyword{internal}
