% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-compilation_engines.R
\name{compile_time_terms_spline}
\alias{compile_time_terms_spline}
\title{Compile time trajectory terms for spline/GAM models}
\usage{
compile_time_terms_spline(time_col, design, k = NULL)
}
\description{
Compile time trajectory terms for spline/GAM models
}
\keyword{internal}
