% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-trajectory_control.R
\name{validate_trajectory_control}
\alias{validate_trajectory_control}
\title{Validate trajectory control arguments}
\usage{
validate_trajectory_control(
  engine,
  family,
  transform,
  na_action,
  spline_k,
  spline_basis,
  spline_method,
  gp_kernel,
  gp_basis_k,
  chains,
  iter,
  cores,
  adapt_delta,
  max_treedepth,
  seed,
  sparse_min_n,
  sparse_other_level,
  missing_level,
  drop_invariant_covariates
)
}
\description{
Validate trajectory control arguments
}
\keyword{internal}
