% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-trajectory_control.R
\name{validate_trajectory_control}
\alias{validate_trajectory_control}
\title{Validate trajectory control arguments}
\usage{
validate_trajectory_control(
  engine,
  family,
  transform,
  na_action,
  spline_k,
  spline_method,
  gp_kernel,
  chains,
  iter,
  cores,
  seed
)
}
\description{
Validate trajectory control arguments
}
\keyword{internal}
